pyproject.toml
[project]
name = "kindle-to-notion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biplist>=1.0.3",
    "google-generativeai>=0.8.5",
    "notion-client>=2.4.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
